修改InformAppAction, MaterialAppAction中AsycTask的执行 更新App测试记录 修复ListView头像乱显示bug 修复红点里显示数字bug 修复version.json乱码bug

+++ core/src/main/java/team/qdu/core/InformAppActionImpl.java
@@ -25,18 +25,26 @@ public class InformAppActionImpl implements InformAppAction {
         this.context = context;
         this.informApi = new InformApiImpl();
     }
+
     //获取当前班课的消息
     @Override
-    public void getInform(final String classId, final ActionCallbackListener<List<Inform>> listener) {
+    public void getInform(final String classId, final Lifeful lifeful, final ActionCallbackListener<List<Inform>> listener) {
         new AsyncTask<Void, Void, ApiResponse<List<Inform>>>() {
 
             @Override
             protected ApiResponse<List<Inform>> doInBackground(Void... params) {
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
                 return informApi.getInform(classId);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<List<Inform>> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(response.getObjList(), response.getMsg());
                 } else {
@@ -47,16 +55,23 @@ public class InformAppActionImpl implements InformAppAction {
     }
 
     @Override
-    public void getUnreadNum(final String informid, final ActionCallbackListener<Void> listener) {
+    public void getUnreadNum(final String informid, final Lifeful lifeful, final ActionCallbackListener<Void> listener) {
         new AsyncTask<Void, Void, ApiResponse<Void>>() {
 
             @Override
             protected ApiResponse<Void> doInBackground(Void... params) {
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
                 return informApi.getUnreadNum(informid);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<Void> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(null, response.getMsg());
                 } else {
@@ -67,16 +82,23 @@ public class InformAppActionImpl implements InformAppAction {
     }
 
     @Override
-    public void getUserInform(final String classId,final String userId, final ActionCallbackListener<List<Inform_User>> listener) {
+    public void getUserInform(final String classId, final String userId, final Lifeful lifeful, final ActionCallbackListener<List<Inform_User>> listener) {
         new AsyncTask<Void, Void, ApiResponse<List<Inform_User>>>() {
 
             @Override
             protected ApiResponse<List<Inform_User>> doInBackground(Void... params) {
-                return informApi.getUserInform(classId,userId);
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
+                return informApi.getUserInform(classId, userId);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<List<Inform_User>> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(response.getObjList(), response.getMsg());
                 } else {
@@ -87,16 +109,23 @@ public class InformAppActionImpl implements InformAppAction {
     }
 
     @Override
-    public void ClickInform(final String inform_user_id,final String classid,final String userid,  final ActionCallbackListener<Void> listener) {
+    public void ClickInform(final String inform_user_id, final String classid, final String userid, final Lifeful lifeful, final ActionCallbackListener<Void> listener) {
         new AsyncTask<Void, Void, ApiResponse<Void>>() {
 
             @Override
             protected ApiResponse<Void> doInBackground(Void... params) {
-                return informApi.ClickInform(inform_user_id,classid,userid);
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
+                return informApi.ClickInform(inform_user_id, classid, userid);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<Void> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(null, response.getMsg());
                 } else {
@@ -108,16 +137,23 @@ public class InformAppActionImpl implements InformAppAction {
     }
 
     @Override
-    public void deleteInform(final String inform_id, final ActionCallbackListener<Void> listener) {
+    public void deleteInform(final String inform_id, final Lifeful lifeful, final ActionCallbackListener<Void> listener) {
         new AsyncTask<Void, Void, ApiResponse<Void>>() {
 
             @Override
             protected ApiResponse<Void> doInBackground(Void... params) {
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
                 return informApi.deleteInform(inform_id);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<Void> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(null, response.getMsg());
                 } else {
@@ -129,16 +165,23 @@ public class InformAppActionImpl implements InformAppAction {
     }
 
     @Override
-    public void getReadPeople(final String inform_id, final ActionCallbackListener<List<User>> listener) {
+    public void getReadPeople(final String inform_id, final Lifeful lifeful, final ActionCallbackListener<List<User>> listener) {
         new AsyncTask<Void, Void, ApiResponse<List<User>>>() {
 
             @Override
             protected ApiResponse<List<User>> doInBackground(Void... params) {
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
                 return informApi.getReadPeople(inform_id);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<List<User>> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(response.getObjList(), response.getMsg());
                 } else {
@@ -149,16 +192,23 @@ public class InformAppActionImpl implements InformAppAction {
     }
 
     @Override
-    public void getUnReadPeople(final String inform_id, final ActionCallbackListener<List<User>> listener) {
+    public void getUnReadPeople(final String inform_id, final Lifeful lifeful, final ActionCallbackListener<List<User>> listener) {
         new AsyncTask<Void, Void, ApiResponse<List<User>>>() {
 
             @Override
             protected ApiResponse<List<User>> doInBackground(Void... params) {
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
                 return informApi.getUnReadPeople(inform_id);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<List<User>> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(response.getObjList(), response.getMsg());
                 } else {
@@ -171,16 +221,23 @@ public class InformAppActionImpl implements InformAppAction {
 
 
     @Override
-    public void createInform(final String classid,final String detail,final ActionCallbackListener<Void> listener){
+    public void createInform(final String classid, final String detail, final Lifeful lifeful, final ActionCallbackListener<Void> listener) {
         new AsyncTask<Void, Void, ApiResponse<Void>>() {
 
             @Override
             protected ApiResponse<Void> doInBackground(Void... params) {
-                return informApi.createInform(classid,detail);
+                if (!lifeful.isAlive()) {
+                    cancel(true);
+                    return null;
+                }
+                return informApi.createInform(classid, detail);
             }
 
             @Override
             protected void onPostExecute(ApiResponse<Void> response) {
+                if (!lifeful.isAlive()) {
+                    return;
+                }
                 if (response.isSuccess()) {
                     listener.onSuccess(null, response.getMsg());
                 } else {
@@ -188,6 +245,5 @@ public class InformAppActionImpl implements InformAppAction {
                 }
             }
         }.execute();
-
     }
 }
